whisper_service/pyproject.toml
[project]
name = "whisper-service"
version = "0.1.0"
description = "Multi-language translation service with Whisper STT"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "redis[hiredis]>=5.0.1",
    "aiohttp>=3.9.1",
    "psutil>=5.9.6",
    "openai>=1.3.7",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "sqlalchemy>=2.0.0",
    "aiomysql>=0.2.0",
    "asyncio-mqtt>=0.13.0",
    "greenlet>=2.0.0",
    "cryptography>=45.0.4",
    "openai-whisper==20240930",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
whisper-service = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# 明确指定要打包的文件
[tool.hatch.build.targets.wheel]
packages = ["services", "utils"]
include = [
    "main.py",
    "config.py",
    "requirements.txt",
    "README.md"
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
